// src/lib/temple/types.ts
export interface TempleState {
  status: TempleStatus;
  accounts: TempleAccount[];
  networks: TempleNetwork[];
  settings: TempleSettings | null;
}

export enum TempleChainId {
  Mainnet = 'NetXdQprcVkpaWU',
  Granadanet = 'NetXz969SFaFn8k',
  Hangzhounet = 'NetXZSsxBpMQeAT',
  Idiazabalnet = 'NetXxkAx4woPLyu'
}

export function isKnownChainId(chainId: string): chainId is TempleChainId {
  return Object.values(TempleChainId).includes(chainId as TempleChainId);
}

export enum TempleStatus {
  Idle,
  Locked,
  Ready
}

export type TempleAccount =
  | TempleHDAccount
  | TempleImportedAccount
  | TempleLedgerAccount
  | TempleManagedKTAccount
  | TempleWatchOnlyAccount;

export enum DerivationType {
  ED25519 = 0,
  SECP256K1 = 1,
  P256 = 2
}

export interface TempleLedgerAccount extends TempleAccountBase {
  type: TempleAccountType.Ledger;
  derivationPath: string;
}

export interface TempleImportedAccount extends TempleAccountBase {
  type: TempleAccountType.Imported;
}

export interface TempleHDAccount extends TempleAccountBase {
  type: TempleAccountType.HD;
  hdIndex: number;
}

export interface TempleManagedKTAccount extends TempleAccountBase {
  type: TempleAccountType.ManagedKT;
  chainId: string;
  owner: string;
}

export interface TempleWatchOnlyAccount extends TempleAccountBase {
  type: TempleAccountType.WatchOnly;
  chainId?: string;
}

export interface TempleAccountBase {
  type: TempleAccountType;
  name: string;
  publicKeyHash: string;
  isActivated?: boolean;
  hdIndex?: number;
  derivationPath?: string;
  derivationType?: DerivationType;
}

export enum TempleAccountType {
  HD,
  Imported,
  Ledger,
  ManagedKT,
  WatchOnly
}

export interface TempleNetwork {
  id: string;
  name: string;
  nameI18nKey?: string;
  description: string;
  descriptionI18nKey?: string;
  type: TempleNetworkType;
  rpcBaseURL: string;
  color: string;
  disabled: boolean;
  hidden?: boolean;
}

export type TempleNetworkType = 'main' | 'test';

export interface TempleSettings {
  customNetworks?: TempleNetwork[];
  lambdaContracts?: Record<string, string>;
  contacts?: TempleContact[];
}

export enum TempleSharedStorageKey {
  DAppEnabled = 'dappenabled',
  LocaleCode = 'localecode',
  UseLedgerLive = 'useledgerlive',
  PasswordAttempts = 'passwordAttempts',
  TimeLock = 'timelock'
}

export type TempleDAppSessions = Record<string, TempleDAppSession>;

export interface TempleContact {
  address: string;
  name: string;
  addedAt?: number;
  accountInWallet?: boolean;
}
// ...
export enum TempleMessageType {
  // Notifications
  StateUpdated = 'TEMPLE_STATE_UPDATED',
  ConfirmationRequested = 'TEMPLE_CONFIRMATION_REQUESTED',
  ConfirmationExpired = 'TEMPLE_CONFIRMATION_EXPIRED',
  // Request-Response pairs
  GetStateRequest = 'TEMPLE_GET_STATE_REQUEST',
  GetStateResponse = 'TEMPLE_GET_STATE_RESPONSE',
  GetSignumTxKeysRequest = 'TEMPLE_GET_SIGNUM_TX_KEYS_REQUEST',
  GetSignumTxKeysResponse = 'TEMPLE_GET_SIGNUM_TX_KEYS_RESPONSE',
  NewWalletRequest = 'TEMPLE_NEW_WALLET_REQUEST',
  NewWalletResponse = 'TEMPLE_NEW_WALLET_RESPONSE',
  UnlockRequest = 'TEMPLE_UNLOCK_REQUEST',
  UnlockResponse = 'TEMPLE_UNLOCK_RESPONSE',
  LockRequest = 'TEMPLE_LOCK_REQUEST',
  LockResponse = 'TEMPLE_LOCK_RESPONSE',
  CreateAccountRequest = 'TEMPLE_CREATE_ACCOUNT_REQUEST',
  CreateAccountResponse = 'TEMPLE_CREATE_ACCOUNT_RESPONSE',
  RevealPublicKeyRequest = 'TEMPLE_REVEAL_PUBLIC_KEY_REQUEST',
  RevealPublicKeyResponse = 'TEMPLE_REVEAL_PUBLIC_KEY_RESPONSE',
  RevealPrivateKeyRequest = 'TEMPLE_REVEAL_PRIVATE_KEY_REQUEST',
  RevealPrivateKeyResponse = 'TEMPLE_REVEAL_PRIVATE_KEY_RESPONSE',
  RevealMnemonicRequest = 'TEMPLE_REVEAL_MNEMONIC_REQUEST',
  RevealMnemonicResponse = 'TEMPLE_REVEAL_MNEMONIC_RESPONSE',
  RemoveAccountRequest = 'TEMPLE_REMOVE_ACCOUNT_REQUEST',
  RemoveAccountResponse = 'TEMPLE_REMOVE_ACCOUNT_RESPONSE',
  ActivateAccountRequest = 'TEMPLE_SET_ACCOUNT_ACTIVATED_REQUEST',
  ActivateAccountResponse = 'TEMPLE_SET_ACCOUNT_ACTIVATED_RESPONSE',
  EditAccountRequest = 'TEMPLE_EDIT_ACCOUNT_REQUEST',
  EditAccountResponse = 'TEMPLE_EDIT_ACCOUNT_RESPONSE',
  ImportAccountRequest = 'TEMPLE_IMPORT_ACCOUNT_REQUEST',
  ImportAccountResponse = 'TEMPLE_IMPORT_ACCOUNT_RESPONSE',
  ImportMnemonicAccountRequest = 'TEMPLE_IMPORT_MNEMONIC_ACCOUNT_REQUEST',
  ImportMnemonicAccountResponse = 'TEMPLE_IMPORT_MNEMONIC_ACCOUNT_RESPONSE',
  ImportFundraiserAccountRequest = 'TEMPLE_IMPORT_FUNDRAISER_ACCOUNT_REQUEST',
  ImportFundraiserAccountResponse = 'TEMPLE_IMPORT_FUNDRAISER_ACCOUNT_RESPONSE',
  ImportManagedKTAccountRequest = 'TEMPLE_IMPORT_MANAGED_KT_ACCOUNT_REQUEST',
  ImportManagedKTAccountResponse = 'TEMPLE_IMPORT_MANAGED_KT_ACCOUNT_RESPONSE',
  ImportWatchOnlyAccountRequest = 'TEMPLE_IMPORT_WATCH_ONLY_ACCOUNT_REQUEST',
  ImportWatchOnlyAccountResponse = 'TEMPLE_IMPORT_WATCH_ONLY_ACCOUNT_RESPONSE',
  CreateLedgerAccountRequest = 'TEMPLE_CREATE_LEDGER_ACCOUNT_REQUEST',
  CreateLedgerAccountResponse = 'TEMPLE_CREATE_LEDGER_ACCOUNT_RESPONSE',
  UpdateSettingsRequest = 'TEMPLE_UPDATE_SETTINGS_REQUEST',
  UpdateSettingsResponse = 'TEMPLE_UPDATE_SETTINGS_RESPONSE',
  OperationsRequest = 'TEMPLE_OPERATIONS_REQUEST',
  OperationsResponse = 'TEMPLE_OPERATIONS_RESPONSE',
  SignRequest = 'TEMPLE_SIGN_REQUEST',
  SignResponse = 'TEMPLE_SIGN_RESPONSE',
  ConfirmationRequest = 'TEMPLE_CONFIRMATION_REQUEST',
  ConfirmationResponse = 'TEMPLE_CONFIRMATION_RESPONSE',
  PageRequest = 'PAGE_REQUEST',
  PageResponse = 'PAGE_RESPONSE',
  DAppGetPayloadRequest = 'TEMPLE_DAPP_GET_PAYLOAD_REQUEST',
  DAppGetPayloadResponse = 'TEMPLE_DAPP_GET_PAYLOAD_RESPONSE',
  DAppPermConfirmationRequest = 'TEMPLE_DAPP_PERM_CONFIRMATION_REQUEST',
  DAppPermConfirmationResponse = 'TEMPLE_DAPP_PERM_CONFIRMATION_RESPONSE',
  DAppOpsConfirmationRequest = 'TEMPLE_DAPP_OPS_CONFIRMATION_REQUEST',
  DAppOpsConfirmationResponse = 'TEMPLE_DAPP_OPS_CONFIRMATION_RESPONSE',
  DAppSignConfirmationRequest = 'TEMPLE_DAPP_SIGN_CONFIRMATION_REQUEST',
  DAppSignConfirmationResponse = 'TEMPLE_DAPP_SIGN_CONFIRMATION_RESPONSE',
  DAppGetAllSessionsRequest = 'TEMPLE_DAPP_GET_ALL_SESSIONS_REQUEST',
  DAppGetAllSessionsResponse = 'TEMPLE_DAPP_GET_ALL_SESSIONS_RESPONSE',
  DAppRemoveSessionRequest = 'TEMPLE_DAPP_REMOVE_SESSION_REQUEST',
  DAppRemoveSessionResponse = 'TEMPLE_DAPP_REMOVE_SESSION_RESPONSE'
}

export type TempleNotification = TempleStateUpdated | TempleConfirmationRequested | TempleConfirmationExpired;

export type TempleRequest =
  | TempleGetStateRequest
  | TempleGetSignumTxKeysRequest
  | TempleNewWalletRequest
  | TempleUnlockRequest
  | TempleLockRequest
  | TempleCreateAccountRequest
  | TempleRevealPublicKeyRequest
  | TempleRevealPrivateKeyRequest
  | TempleRevealMnemonicRequest
  | TempleActivateAccountRequest
  | TempleEditAccountRequest
  | TempleImportAccountRequest
  | TempleImportMnemonicAccountRequest
  | TempleImportFundraiserAccountRequest
  | TempleImportManagedKTAccountRequest
  | TempleImportWatchOnlyAccountRequest
  | TempleCreateLedgerAccountRequest
  | TempleOperationsRequest
  | TempleSignRequest
  | TempleConfirmationRequest
  | TempleRemoveAccountRequest
  | TemplePageRequest
  | TempleDAppGetPayloadRequest
  | TempleDAppPermConfirmationRequest
  | TempleDAppOpsConfirmationRequest
  | TempleDAppSignConfirmationRequest
  | TempleUpdateSettingsRequest
  | TempleGetAllDAppSessionsRequest
  | TempleRemoveDAppSessionRequest;
// ...
export interface TempleMessageBase {
  type: TempleMessageType;
}
// ...
export interface TempleGetStateRequest extends TempleMessageBase {
  type: TempleMessageType.GetStateRequest;
}

export interface TempleGetStateResponse extends TempleMessageBase {
  type: TempleMessageType.GetStateResponse;
  state: TempleState;
}

export interface TempleGetSignumTxKeysRequest extends TempleMessageBase {
  type: TempleMessageType.GetSignumTxKeysRequest;
  accountPublicKeyHash: string;
}

export interface TempleGetSignumTxKeysResponse extends TempleMessageBase {
  type: TempleMessageType.GetSignumTxKeysResponse;
  publicKey: string;
  signingKey: string;
}

export interface TempleNewWalletRequest extends TempleMessageBase {
  type: TempleMessageType.NewWalletRequest;
  password: string;
  mnemonic?: string;
}

export interface TempleNewWalletResponse extends TempleMessageBase {
  type: TempleMessageType.NewWalletResponse;
}

export interface TempleUnlockRequest extends TempleMessageBase {
  type: TempleMessageType.UnlockRequest;
  password: string;
}

export interface TempleUnlockResponse extends TempleMessageBase {
  type: TempleMessageType.UnlockResponse;
}

export interface TempleLockRequest extends TempleMessageBase {
  type: TempleMessageType.LockRequest;
}

export interface TempleLockResponse extends TempleMessageBase {
  type: TempleMessageType.LockResponse;
}

export interface TempleCreateAccountRequest extends TempleMessageBase {
  type: TempleMessageType.CreateAccountRequest;
  name?: string;
}

export interface TempleCreateAccountResponse extends TempleMessageBase {
  type: TempleMessageType.CreateAccountResponse;
  mnemonic: string;
}

export interface TempleRevealPublicKeyRequest extends TempleMessageBase {
  type: TempleMessageType.RevealPublicKeyRequest;
  accountPublicKeyHash: string;
}

export interface TempleRevealPublicKeyResponse extends TempleMessageBase {
  type: TempleMessageType.RevealPublicKeyResponse;
  publicKey: string;
}

export interface TempleRevealPrivateKeyRequest extends TempleMessageBase {
  type: TempleMessageType.RevealPrivateKeyRequest;
  accountPublicKeyHash: string;
  password: string;
}

export interface TempleRevealPrivateKeyResponse extends TempleMessageBase {
  type: TempleMessageType.RevealPrivateKeyResponse;
  privateKey: string;
}

export interface TempleRevealMnemonicRequest extends TempleMessageBase {
  type: TempleMessageType.RevealMnemonicRequest;
  password: string;
}

export interface TempleRevealMnemonicResponse extends TempleMessageBase {
  type: TempleMessageType.RevealMnemonicResponse;
  mnemonic: string;
}

export interface TempleRemoveAccountRequest extends TempleMessageBase {
  type: TempleMessageType.RemoveAccountRequest;
  accountPublicKeyHash: string;
  password: string;
}

export interface TempleRemoveAccountResponse extends TempleMessageBase {
  type: TempleMessageType.RemoveAccountResponse;
}

export interface TempleActivateAccountRequest extends TempleMessageBase {
  type: TempleMessageType.ActivateAccountRequest;
  accountPublicKeyHash: string;
}

export interface TempleActivateAccountResponse extends TempleMessageBase {
  type: TempleMessageType.ActivateAccountResponse;
}

export interface TempleEditAccountRequest extends TempleMessageBase {
  type: TempleMessageType.EditAccountRequest;
  accountPublicKeyHash: string;
  name: string;
}

export interface TempleEditAccountResponse extends TempleMessageBase {
  type: TempleMessageType.EditAccountResponse;
}

export interface TempleImportAccountRequest extends TempleMessageBase {
  type: TempleMessageType.ImportAccountRequest;
  privateKey: string;
  encPassword?: string;
}

export interface TempleImportAccountResponse extends TempleMessageBase {
  type: TempleMessageType.ImportAccountResponse;
}

export interface TempleImportMnemonicAccountRequest extends TempleMessageBase {
  type: TempleMessageType.ImportMnemonicAccountRequest;
  mnemonic: string;
  name?: string;
}

export interface TempleImportMnemonicAccountResponse extends TempleMessageBase {
  type: TempleMessageType.ImportMnemonicAccountResponse;
}

export interface TempleImportFundraiserAccountRequest extends TempleMessageBase {
  type: TempleMessageType.ImportFundraiserAccountRequest;
  email: string;
  password: string;
  mnemonic: string;
}

export interface TempleImportFundraiserAccountResponse extends TempleMessageBase {
  type: TempleMessageType.ImportFundraiserAccountResponse;
}

export interface TempleImportManagedKTAccountRequest extends TempleMessageBase {
  type: TempleMessageType.ImportManagedKTAccountRequest;
  address: string;
  chainId: string;
  owner: string;
}

export interface TempleImportManagedKTAccountResponse extends TempleMessageBase {
  type: TempleMessageType.ImportManagedKTAccountResponse;
}

export interface TempleImportWatchOnlyAccountRequest extends TempleMessageBase {
  type: TempleMessageType.ImportWatchOnlyAccountRequest;
  address: string;
  chainId?: string;
}

export interface TempleImportWatchOnlyAccountResponse extends TempleMessageBase {
  type: TempleMessageType.ImportWatchOnlyAccountResponse;
}

export interface TempleCreateLedgerAccountRequest extends TempleMessageBase {
  type: TempleMessageType.CreateLedgerAccountRequest;
  name: string;
  derivationPath?: string;
  derivationType?: DerivationType;
}

export interface TempleCreateLedgerAccountResponse extends TempleMessageBase {
  type: TempleMessageType.CreateLedgerAccountResponse;
}

export interface TempleUpdateSettingsRequest extends TempleMessageBase {
  type: TempleMessageType.UpdateSettingsRequest;
  settings: Partial<TempleSettings>;
}

export interface TempleUpdateSettingsResponse extends TempleMessageBase {
  type: TempleMessageType.UpdateSettingsResponse;
}

export interface TempleOperationsRequest extends TempleMessageBase {
  type: TempleMessageType.OperationsRequest;
  id: string;
  sourcePkh: string;
  networkRpc: string;
  opParams: any[];
}

export interface TempleOperationsResponse extends TempleMessageBase {
  type: TempleMessageType.OperationsResponse;
  opHash: string;
}

export interface TempleSignRequest extends TempleMessageBase {
  type: TempleMessageType.SignRequest;
  id: string;
  sourcePkh: string;
  bytes: string;
  watermark?: string;
}

export interface TempleSignResponse extends TempleMessageBase {
  type: TempleMessageType.SignResponse;
  result: any;
}

export interface TempleConfirmationRequest extends TempleMessageBase {
  type: TempleMessageType.ConfirmationRequest;
  id: string;
  confirmed: boolean;
  modifiedTotalFee?: number;
  modifiedStorageLimit?: number;
}

export interface TempleConfirmationResponse extends TempleMessageBase {
  type: TempleMessageType.ConfirmationResponse;
}

export interface TemplePageRequest extends TempleMessageBase {
  type: TempleMessageType.PageRequest;
  origin: string;
  payload: any;
  beacon?: boolean;
  encrypted?: boolean;
}

export interface TemplePageResponse extends TempleMessageBase {
  type: TempleMessageType.PageResponse;
  payload: any;
  encrypted?: boolean;
}

export interface TempleDAppGetPayloadRequest extends TempleMessageBase {
  type: TempleMessageType.DAppGetPayloadRequest;
  id: string;
}

export interface TempleDAppGetPayloadResponse extends TempleMessageBase {
  type: TempleMessageType.DAppGetPayloadResponse;
  payload: TempleDAppPayload;
}

export interface TempleDAppPermConfirmationRequest extends TempleMessageBase {
  type: TempleMessageType.DAppPermConfirmationRequest;
  id: string;
  confirmed: boolean;
  accountPublicKey: string;
  accountPublicKeyHash: string;
}

export interface TempleDAppPermConfirmationResponse extends TempleMessageBase {
  type: TempleMessageType.DAppPermConfirmationResponse;
}

export interface TempleDAppOpsConfirmationRequest extends TempleMessageBase {
  type: TempleMessageType.DAppOpsConfirmationRequest;
  id: string;
  confirmed: boolean;
  modifiedTotalFee?: number;
  modifiedStorageLimit?: number;
}

export interface TempleDAppOpsConfirmationResponse extends TempleMessageBase {
  type: TempleMessageType.DAppOpsConfirmationResponse;
}

export interface TempleDAppSignConfirmationRequest extends TempleMessageBase {
  type: TempleMessageType.DAppSignConfirmationRequest;
  id: string;
  confirmed: boolean;
}

export interface TempleDAppSignConfirmationResponse extends TempleMessageBase {
  type: TempleMessageType.DAppSignConfirmationResponse;
}

export interface TempleGetAllDAppSessionsRequest extends TempleMessageBase {
  type: TempleMessageType.DAppGetAllSessionsRequest;
}

export interface TempleGetAllDAppSessionsResponse extends TempleMessageBase {
  type: TempleMessageType.DAppGetAllSessionsResponse;
  sessions: TempleDAppSessions;
}

export interface TempleRemoveDAppSessionRequest extends TempleMessageBase {
  type: TempleMessageType.DAppRemoveSessionRequest;
  origin: string;
}